+++ module-gtnhmixins/src/main/java/com/gtnewhorizon/gtnhmixins/builders/ITargetMod.java
@@ -1,4 +1,4 @@
-package com.gtnewhorizon.gtnhmixins.mixinbuilder;
+package com.gtnewhorizon.gtnhmixins.builders;
 
 import org.objectweb.asm.tree.ClassNode;
